refactor: restructure Subreddit module
Subreddit now has two subpackages: api & impl. Code is currently in impl.
