+++ app/src/main/kotlin/com/potatosheep/kite/app/nav/KiteNavHost.kt
@@ -18,8 +18,8 @@ import com.potatosheep.kite.feature.post.impl.nav.postScreen
 import com.potatosheep.kite.feature.search.impl.nav.navigateToSearch
 import com.potatosheep.kite.feature.search.impl.nav.searchScreen
 import com.potatosheep.kite.feature.settings.impl.nav.settingsScreen
-import com.potatosheep.kite.feature.subreddit.nav.navigateToSubreddit
-import com.potatosheep.kite.feature.subreddit.nav.subredditScreen
+import com.potatosheep.kite.feature.subreddit.impl.nav.navigateToSubreddit
+import com.potatosheep.kite.feature.subreddit.impl.nav.subredditScreen
 import com.potatosheep.kite.feature.user.nav.navigateToUser
 import com.potatosheep.kite.feature.user.nav.userScreen
 import com.potatosheep.kite.feature.video.nav.navigateToVideo
