+++ feature/subreddit/impl/src/main/AndroidManifest.xml
