+++ feature/subreddit/impl/src/main/kotlin/com/potatosheep/kite/feature/subreddit/impl/SubredditPostContent.kt
@@ -1,4 +1,4 @@
-package com.potatosheep.kite.feature.subreddit
+package com.potatosheep.kite.feature.subreddit.impl
 
 import androidx.compose.animation.AnimatedVisibilityScope
 import androidx.compose.animation.ExperimentalSharedTransitionApi
