+++ feature/subreddit/impl/src/main/res/drawable/ic_launcher_background.xml
