+++ feature/subreddit/impl/src/main/res/mipmap-xxhdpi/ic_launcher_round.webp
