+++ settings.gradle.kts
@@ -39,7 +39,8 @@ include(":feature:post:impl")
 include(":feature:post:api")
 include(":feature:image:impl")
 include(":feature:image:api")
-include(":feature:subreddit")
+include(":feature:subreddit:impl")
+include(":feature:subreddit:api")
 include(":core:common")
 include(":feature:subredditwiki")
 include(":feature:user")
